break out recovery tests from reindex spec in e2e
