+++ e2e/test/cases/data/reindex-spec.js
@@ -1,7 +1,6 @@
 'use strict';
 
 const _ = require('lodash');
-const Promise = require('bluebird');
 const misc = require('../../misc');
 const { waitForJobStatus } = require('../../wait');
 const { resetState, testJobLifeCycle, runEsJob } = require('../../helpers');
@@ -72,26 +71,6 @@ describe('reindex', () => {
 
         const stats = await misc.indexStats(specIndex);
         expect(stats.count).toBe(1000);
-    });
-
-    it('can support different recovery mode cleanup=errors', async () => {
-        const errorStates = '/ex/testex-errors/_recover?cleanup=errors';
-
-        const { job_id: jobId } = await teraslice.cluster.post(errorStates);
-        const job = teraslice.jobs.wrap(jobId);
-        await waitForJobStatus(job, 'completed');
-        const stats = await misc.indexStats('test-recovery-100');
-        expect(stats.count).toEqual(100);
-    });
-
-    it('can support different recovery mode cleanup=all', async () => {
-        const allStates = '/ex/testex-all/_recover?cleanup=all';
-
-        const { job_id: jobId } = await teraslice.cluster.post(allStates);
-        const job = teraslice.jobs.wrap(jobId);
-        await waitForJobStatus(job, 'completed');
-        const stats = await misc.indexStats('test-recovery-200');
-        expect(stats.count).toEqual(200);
     });
 
     it('should support idempotency', async () => {
